# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FinEdge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/me:
    get:
      operationId: getMe
      tags: [user]
      summary: Get current user profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [user]
      summary: Update user profile
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /user/quiz:
    get:
      operationId: getQuizPreferences
      tags: [user]
      summary: Get user trading quiz preferences
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Quiz preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizPreferences"
    post:
      operationId: saveQuizPreferences
      tags: [user]
      summary: Save user trading quiz preferences
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizPreferencesInput"
      responses:
        "200":
          description: Saved quiz preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizPreferences"
  /user/streak:
    get:
      operationId: getLearningStreak
      tags: [user]
      summary: Get user learning streak data
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Streak data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningStreak"
    post:
      operationId: checkInStreak
      tags: [user]
      summary: Record daily check-in for streak
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Updated streak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningStreak"
...
  /stocks/discover:
    get:
      operationId: discoverStocks
      tags: [stocks]
      summary: Get curated stock discovery list based on user preferences
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Discovered stocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockCard"
  /stocks/quote/{ticker}:
    get:
      operationId: getStockQuote
      tags: [stocks]
      summary: Get current stock quote
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stock quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockQuote"
  /stocks/ohlc/{ticker}:
    get:
      operationId: getStockOHLC
      tags: [stocks]
      summary: Get OHLC candle data for a stock (last 5 days)
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OHLC candle data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OHLCCandle"
  /stocks/search:
    get:
      operationId: searchStocks
      tags: [stocks]
      summary: Search stocks by ticker or company name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockSearchResult"
  /stocks/insight:
    get:
      operationId: getMarketInsights
      tags: [stocks]
      summary: Get daily buy signal and slowdown alert
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Market insights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketInsights"
...
  /stripe/subscription:
    get:
      operationId: getSubscription
      tags: [stripe]
      summary: Get current user subscription status
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Subscription info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
  /stripe/checkout:
    post:
      operationId: createCheckoutSession
      tags: [stripe]
      summary: Create Stripe checkout session for Pro subscription
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
  /stripe/portal:
    post:
      operationId: createPortalSession
      tags: [stripe]
      summary: Create Stripe customer portal session
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Portal session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalSession"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserProfile:
      type: object
      required: [id, email, isPro]
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        isPro:
          type: boolean
        stripeCustomerId:
          type: ["string", "null"]
        stripeSubscriptionId:
          type: ["string", "null"]
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
    QuizPreferences:
      type: object
      required: [userId, completed]
      properties:
        userId:
          type: string
        strategies:
          type: array
          items:
            type: string
        capitalRange:
          type: ["string", "null"]
        profitTarget:
          type: ["string", "null"]
        completed:
          type: boolean
    QuizPreferencesInput:
      type: object
      required: [strategies, capitalRange, profitTarget]
      properties:
        strategies:
          type: array
          items:
            type: string
        capitalRange:
          type: string
        profitTarget:
          type: string
    LearningStreak:
      type: object
      required: [currentStreak, longestStreak, lastCheckinDate, checkinDates]
      properties:
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        lastCheckinDate:
          type: ["string", "null"]
        checkinDates:
          type: array
          items:
            type: string
...
    StockQuote:
      type: object
      required: [ticker, price, change, changePercent, previousClose]
      properties:
        ticker:
          type: string
        companyName:
          type: ["string", "null"]
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        previousClose:
          type: number
        previousHigh:
          type: number
        previousLow:
          type: number
        volume:
          type: ["integer", "null"]
        avgVolume:
          type: ["integer", "null"]
        atr:
          type: ["number", "null"]
        isHighVolatility:
          type: boolean
        sparkline:
          type: array
          items:
            type: number
    StockCard:
      type: object
      required: [ticker, companyName, price, changePercent, previousHigh, previousLow, candles]
      properties:
        ticker:
          type: string
        companyName:
          type: string
        price:
          type: number
        changePercent:
          type: number
        previousHigh:
          type: number
        previousLow:
          type: number
        atr:
          type: ["number", "null"]
        isHighVolatility:
          type: boolean
        candles:
          type: array
          items:
            $ref: "#/components/schemas/OHLCCandle"
    OHLCCandle:
      type: object
      required: [time, open, high, low, close, volume]
      properties:
        time:
          type: string
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: integer
    StockSearchResult:
      type: object
      required: [ticker, companyName]
      properties:
        ticker:
          type: string
        companyName:
          type: string
        sector:
          type: ["string", "null"]
    MarketInsights:
      type: object
      required: [opportunity, slowdown]
      properties:
        opportunity:
          $ref: "#/components/schemas/InsightCard"
        slowdown:
          $ref: "#/components/schemas/InsightCard"
    InsightCard:
      type: object
      required: [ticker, companyName, summary, label, generatedAt]
      properties:
        ticker:
          type: string
        companyName:
          type: string
        summary:
          type: string
        label:
          type: string
        changePercent30d:
          type: ["number", "null"]
        generatedAt:
          type: string
...
    SubscriptionStatus:
      type: object
      required: [isPro, status]
      properties:
        isPro:
          type: boolean
        status:
          type: ["string", "null"]
        currentPeriodEnd:
          type: ["string", "null"]
        priceId:
          type: ["string", "null"]
    CheckoutInput:
      type: object
      required: [priceId]
      properties:
        priceId:
          type: string
    CheckoutSession:
      type: object
      required: [url]
      properties:
        url:
          type: string
    PortalSession:
      type: object
      required: [url]
      properties:
        url:
          type: string
...