# lib/api-zod/src/generated/types/quizPreferencesInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FinEdge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizPreferencesInput {
  strategies: string[];
  capitalRange: string;
  profitTarget: string;
}