# lib/api-zod/src/generated/types/stockCard.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FinEdge API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OHLCCandle } from './oHLCCandle';

export interface StockCard {
  ticker: string;
  companyName: string;
  price: number;
  changePercent: number;
  previousHigh: number;
  previousLow: number;
  /** @nullable */
  atr?: number | null;
  isHighVolatility?: boolean;
  candles: OHLCCandle[];
}